removed short type usages Rationale: Memory savings are not guaranteed (compiler and block size dependent), but definitely needs higher run time for casting forth and back to int.

+++ trunk/source/Library-Automata/src/de/uni_freiburg/informatik/ultimate/automata/nestedword/StateContainerFieldMap.java
@@ -188,7 +188,7 @@ public class StateContainerFieldMap<LETTER, STATE> {
 		/**
 		 * Points to next field with {@link OutgoingInternalTransition}.
 		 */
-		private short mPosition;
+		private int mPosition;
 		
 		public FieldIterator() {
 			updatePosition();
